Refactored BettingHandler package. Add Profession command to change the outfits of the NPCs. Minor code cleanup. Completely remove tab completion as it doesn't always work. Make commands check for more arguments rather than exact number of arguments. Fix some unhandled exceptions in EntityListener. Added the ability to add custom messages to KitNPC's when they are interacted with. Shorts and Double aren't the same thing.

+++ EconomyNPC V2.0/src/com/mongoosecountry/pete800/util/Prices.java
@@ -64,7 +64,7 @@ public class Prices
 				String material = entry.getKey().split("\\.")[0].toUpperCase();
 				short durability = Short.valueOf(entry.getKey().split("\\.")[1]);
 				ItemStack item = new ItemStack(Material.getMaterial(material), 0, durability);
-				double price = Short.valueOf(entry.getValue().toString());
+				double price = Double.valueOf(entry.getValue().toString());
 				prices.put(item, price);
 			}
 		}
